move remaining scss off pages and components
