+++ src/docs/shared/components/CarbonBadge.tsx
@@ -1,8 +1,7 @@
 // TODO: ADS-153 Do we want to add internationalization to this component and include it in the library?
 
 import { useEffect, useState } from 'react';
-import { getStorage, setStorage } from '../../helpers';
-import './CarbonBadge.scss';
+import { getStorage, setStorage } from '../helpers';
 
 interface CarbonData {
   c: number; // Carbon amount
