+++ src/docs/styles/components/_components.scss
@@ -1,0 +1,1 @@
+@import 'carbon-badge';
