+++ src/docs/styles/global.scss
@@ -3,3 +3,5 @@
 @import './layout/layout';
 @import './themes/markdown-theme'; // theme for code snippets
 @import './docs/docs';
+@import './components/components';
+@import './pages/pages';
