+++ src/docs/styles/pages/_home.scss
@@ -1,9 +1,3 @@
-// TODO: Remove import once stylesheet is moved off of component
-@import '~@boston-scientific/anatomy-tokens/lib/sass/tokens';
-
-@import '/src/library/styles/abstracts/abstracts';
-@import '/src/library/styles/base/base';
-
 .docs-home {
   position: relative;
   display: grid;
