+++ src/docs/styles/pages/_pages.scss
@@ -1,0 +1,2 @@
+@import 'home';
+@import 'search-results';
