+++ src/docs/styles/pages/_search-results.scss
@@ -1,9 +1,3 @@
-// TODO: Remove import once stylesheet is moved off of component
-@import '~@boston-scientific/anatomy-tokens/lib/sass/tokens';
-
-@import '/src/library/styles/abstracts/abstracts';
-@import '/src/library/styles/base/base';
-
 .docs-search-results {
   margin: 0;
   margin-top: space(6x);
